Added weapon stats box, updated database for enhance levels.
